+++ onomaturgy/__init__.py
@@ -1,0 +1,4 @@
+from onomaturgy.generators.personal_name import PersonalNameGenerator
+from onomaturgy.generators.place_name import PlaceNameGenerator
+from onomaturgy.generators.simple_name import SimpleNameGenerator
+from onomaturgy.generators.tribal_name import TribalNameGenerator